Delete FlyoverRedeemScriptCreationException and move flyover reason to RedeemScriptCreationException
