+++ rskj-core/src/main/java/co/rsk/peg/bitcoin/RedeemScriptCreationException.java
@@ -5,7 +5,8 @@ public class RedeemScriptCreationException extends RuntimeException {
 
     public enum Reason {
         INVALID_INTERNAL_REDEEM_SCRIPTS,
-        INVALID_CSV_VALUE
+        INVALID_CSV_VALUE,
+        INVALID_FLYOVER_DERIVATION_HASH
     }
 
     public RedeemScriptCreationException(String s, Reason reason) {
